code/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pypdf
python-docx
ollama
langchain-text-splitters
numpy
